Added boolean, fixed .gitignore and ammended naming to match convention
