+++ .gitignore
@@ -258,3 +258,13 @@ MidasCivil_Adapter/obj/Debug/MidasCivil_Adapter.dll
 MidasCivil_Adapter/obj/Debug/MidasCivil_Adapter.pdb
 MidasCivil_Engine/obj/Debug/MidasCivil_Engine.dll
 MidasCivil_Engine/obj/Debug/MidasCivil_Engine.pdb
+Build/Common_oM.dll
+Build/Common_oM.pdb
+Build/Geometry_oM.dll
+Build/Geometry_oM.pdb
+Build/MidasCivil_Adapter.dll
+Build/MidasCivil_Adapter.pdb
+Build/MidasCivil_Engine.dll
+Build/MidasCivil_Engine.pdb
+Build/Structure_oM.dll
+Build/Structure_oM.pdb
