+++ MidasCivil_Engine/MidasCivil_Engine.csproj
@@ -71,9 +71,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Convert\AdapterId.cs" />
-    <Compile Include="Convert\PrivateHelpers\bhomTypeConvert.cs" />
+    <Compile Include="Convert\PrivateHelpers\BHoMTypeConvert.cs" />
     <Compile Include="Convert\PrivateHelpers\CheckSupportedDOFType.cs" />
-    <Compile Include="Convert\PrivateHelpers\CompileTextFiles.cs" />
+    <Compile Include="Convert\PrivateHelpers\CombineTextFiles.cs" />
     <Compile Include="Convert\PrivateHelpers\GetSpringStiffness.cs" />
     <Compile Include="Convert\PrivateHelpers\SupportToString.cs" />
     <Compile Include="Convert\ToBHoM\Elements\Bar.cs" />
